requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
et-xmlfile==1.1.0
frozenlist==1.3.1
idna==3.4
meraki==1.24.0
multidict==6.0.2
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
urllib3==1.26.12
yarl==1.8.1
